backend/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
langchain_core==0.3.28
langchain_openai==0.2.14
langgraph==0.2.60
matplotlib==3.10.0
openai
langchain
langchain-openai
youtube-transcript-api
wikipedia==1.4.0
beautifulsoup4
scikit-learn
googlesearch-python
requests
numpy
flask-caching
